requirement.txt
asgiref==3.8.1
coverage==7.8.0
Django==5.2
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-nested-routers==0.94.1
iniconfig==2.1.0
Markdown==3.8
packaging==24.2
pillow==11.2.1
pluggy==1.5.0
PyJWT==2.9.0
pytest==8.3.5
pytest-cov==6.1.1
pytest-django==4.11.1
python-dotenv==1.1.0
sqlparse==0.5.3
